implement very primitive generic functions with inference in return type position
inference in the return type is predicated on analyzing the flow of function bodies,
which really only works in the simplest case right now -- unbraced arrow functions,
where the flow doesn't need to be analyzed because the return type is the type of the
single expression of the body.

While primitive and not useable in this condition, it should work well once we
get control flow analysis somewhat implemented.