rough impl of conditional type based mapped component property injections
supports a syntax like
/**
* @!typedef properties<T> = T.cfname extends string
* ? inject<"addPrefix#T.cfname#", () => any>
* : 0
*
*/

where "properties<T>" is a type the compiler will look for when running
a type check pass over each property; `T` is a type parameter bound to
information about the property when `properties<T>` is instantiated;
`T.cfname` is a lookup into T; `inject<name, type>` is a compiler provided type made
availabe in the context of the evaluation of `properties<T>` and serves to add a binding
with a particular name and type to the variables scope.

we can inject properties, but maybe we want to erase them, too?
also, we need template string literals that can match extends clauses like "a#infer V#c"
which would match like the regex `a(.+)c` and bind type parameter V to the capture group